nlp-service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
spacy==3.8.7
python-dotenv==1.1.0
